Improved LZW Encoder: see description
(mostly in a previous commit) Our main improvement to the encoder was that, when possible, we converted the outputted integers to chars (so when the integer was < 256 because we're using UTF-8). This will improve the compression because chars take up less space than integers. Additionally, we changed the "magic number" of 127 to a constant and added in some more comments.